Make Monaco Editor compatible with Vite
Monaco Editor uses some static properties that reference the class they
are defined in. This is equivalent to `this`. Monaco Editor also uses
decorators. TypeScript compiles this in a way that works (with some
small boilerplate code). In Vite however, this compiles to broken code.

This change replaces self-references via in static properties via their
identifier with `this`.

The changes were cherry-picked from #249616.